Move fontinfo types into shared_types
